Fix error handling and code reuse.
