+++ examples/gas_free/sources/gas_free.move
@@ -1,14 +1,14 @@
 module rooch_examples::gas_payer {
-    use moveos_std::storage_context::StorageContext;
+    use moveos_std::context::Context;
 
-    fun gas_validate_function(_storage_ctx: &StorageContext): bool{
+    fun gas_validate_function(_ctx: &Context): bool{
         true
     }
 
-    fun gas_charge_post_function(_storage_ctx: &mut StorageContext, _gas_used: u128) {
+    fun gas_charge_post_function(_ctx: &mut Context, _gas_used: u128) {
     }
 
     #[gas_free(gas_validate=gas_validate_function, gas_charge_post=gas_charge_post_function)]
-    public entry fun play(_ctx: &mut StorageContext, _sender: &signer){
+    public entry fun play(_ctx: &mut Context, _sender: &signer){
     }
 }
